fix: report latency metric directly from gateway
